feat!: add target flag
This commit adds support for specifying the target triple that blackjack
uses when generating the Bazel rules.

BREAKING CHANGE: The `-t,--target` argument is mandatory. Existing
tooling fails to run successfully due to it missing. To achieve the same
behavior as in v0.1.0, the target value should be set to
`x86_64-unknown-linux-gnu`.

Closes #5